#11, #20 - Add slime king enemy. Refactor animation logic.
BowmanAnimation, ClericIdle, KnightAsset, MageAnimationAsset - Added trigger and integer to switch animation states, instead of play state with string.

SlimeKingReal - A PSB file contains image of Slime King.

SlimeKing, SlimeIdle - Animation assets for Slime King.

AttackMotion - Added trigger and integer to switch animation states, instead of play state with string. Now it supports a function to move characters.

CombatManager - Now cast from MatchTier to integer to call PlayAttackMotion call.

+++ Project_JM/Assets/Scripts/Systems/CombatManager.cs
@@ -113,13 +113,13 @@ public class CombatManager : MonoBehaviour
     {
         if (context.Attacker is MonoBehaviour attackerObject)
         {
-            attackerObject.GetComponent<AttackMotion>().PlayAttackMotion(matchTier);
+            attackerObject.GetComponent<AttackMotion>().PlayAttackMotion((int)matchTier);
         }
 
         // @@ TODO: Play hit animation to target.
         if (context.Target is MonoBehaviour targetObject)
         {
-            targetObject.GetComponent<AttackMotion>().PlayAttackMotion(matchTier);
+            targetObject.GetComponent<AttackMotion>().PlayAttackMotion((int)matchTier);
         }
     }
